add bill module
