+++ src/bill/identity.rs
@@ -1,0 +1,252 @@
+use crate::util;
+use borsh::{to_vec, BorshDeserialize};
+use borsh_derive::{BorshDeserialize, BorshSerialize};
+use libp2p::identity::Keypair;
+use libp2p::PeerId;
+use openssl::pkey::Private;
+use openssl::rsa::Rsa;
+use rocket::serde::{Deserialize, Serialize};
+use rocket::FromForm;
+use std::{fs, mem};
+
+use crate::constants::{
+    IDENTITY_ED_25529_KEYS_FILE_PATH, IDENTITY_FILE_PATH, IDENTITY_PEER_ID_FILE_PATH, USEDNET,
+};
+#[derive(BorshSerialize, BorshDeserialize, FromForm, Debug, Serialize, Deserialize, Clone)]
+#[serde(crate = "rocket::serde")]
+pub struct NodeId {
+    id: String,
+}
+
+impl NodeId {
+    pub fn new(peer_id: String) -> Self {
+        Self { id: peer_id }
+    }
+}
+
+#[derive(Clone)]
+pub struct IdentityWithAll {
+    pub identity: Identity,
+    pub peer_id: PeerId,
+    pub key_pair: Keypair,
+}
+
+#[derive(BorshSerialize, BorshDeserialize, FromForm, Debug, Serialize, Deserialize, Clone)]
+#[serde(crate = "rocket::serde")]
+pub struct Identity {
+    pub name: String,
+    pub company: String,
+    pub date_of_birth: String,
+    pub city_of_birth: String,
+    pub country_of_birth: String,
+    pub email: String,
+    pub postal_address: String,
+    pub public_key_pem: String,
+    pub private_key_pem: String,
+    pub bitcoin_public_key: String,
+    pub bitcoin_private_key: String,
+}
+
+macro_rules! update_field {
+    ($self:expr, $other:expr, $field:ident) => {
+        if !$other.$field.is_empty() {
+            $self.$field = $other.$field.clone();
+        }
+    };
+}
+
+impl Identity {
+    pub fn new_empty() -> Self {
+        Self {
+            name: "".to_string(),
+            company: "".to_string(),
+            date_of_birth: "".to_string(),
+            city_of_birth: "".to_string(),
+            bitcoin_public_key: "".to_string(),
+            postal_address: "".to_string(),
+            public_key_pem: "".to_string(),
+            email: "".to_string(),
+            country_of_birth: "".to_string(),
+            private_key_pem: "".to_string(),
+            bitcoin_private_key: "".to_string(),
+        }
+    }
+
+    fn all_changeable_fields_empty(&self) -> bool {
+        self.name.is_empty()
+            && self.company.is_empty()
+            && self.postal_address.is_empty()
+            && self.email.is_empty()
+    }
+
+    fn all_changeable_fields_equal_to(&self, other: &Self) -> bool {
+        self.name == other.name
+            && self.company == other.company
+            && self.postal_address == other.postal_address
+            && self.email == other.email
+    }
+
+    pub fn update_valid(&self, other: &Self) -> bool {
+        if other.all_changeable_fields_empty() {
+            return false;
+        }
+        if self.all_changeable_fields_equal_to(other) {
+            return false;
+        }
+        true
+    }
+
+    pub fn update_from(&mut self, other: &Identity) {
+        update_field!(self, other, name);
+        update_field!(self, other, company);
+        update_field!(self, other, postal_address);
+        update_field!(self, other, email);
+    }
+}
+
+pub fn get_whole_identity() -> IdentityWithAll {
+    let identity: Identity = read_identity_from_file();
+    let ed25519_keys: Keypair = read_ed25519_keypair_from_file();
+    let peer_id: PeerId = read_peer_id_from_file();
+
+    IdentityWithAll {
+        identity,
+        peer_id,
+        key_pair: ed25519_keys,
+    }
+}
+
+pub fn create_whole_identity(
+    name: String,
+    company: String,
+    date_of_birth: String,
+    city_of_birth: String,
+    country_of_birth: String,
+    email: String,
+    postal_address: String,
+) -> IdentityWithAll {
+    let identity = create_new_identity(
+        name,
+        company,
+        date_of_birth,
+        city_of_birth,
+        country_of_birth,
+        email,
+        postal_address,
+    );
+
+    let ed25519_keys = read_ed25519_keypair_from_file();
+    let peer_id = read_peer_id_from_file();
+
+    write_identity_to_file(&identity);
+
+    IdentityWithAll {
+        identity,
+        peer_id,
+        key_pair: ed25519_keys,
+    }
+}
+
+pub fn generate_dht_logic() {
+    let ed25519_keys = Keypair::generate_ed25519();
+    let peer_id = ed25519_keys.public().to_peer_id();
+
+    write_dht_logic(&peer_id, &ed25519_keys);
+}
+
+fn write_dht_logic(peer_id: &PeerId, ed25519_keys: &Keypair) {
+    write_peer_id_to_file(peer_id);
+    write_ed25519_keypair_to_file(ed25519_keys);
+}
+
+pub fn create_new_identity(
+    name: String,
+    company: String,
+    date_of_birth: String,
+    city_of_birth: String,
+    country_of_birth: String,
+    email: String,
+    postal_address: String,
+) -> Identity {
+    let rsa: Rsa<Private> = util::rsa::generation_rsa_key();
+    let private_key_pem: String = util::rsa::pem_private_key_from_rsa(&rsa);
+    let public_key_pem: String = util::rsa::pem_public_key_from_rsa(&rsa);
+
+    let s = bitcoin::secp256k1::Secp256k1::new();
+    let private_key = bitcoin::PrivateKey::new(
+        s.generate_keypair(&mut bitcoin::secp256k1::rand::thread_rng())
+            .0,
+        USEDNET,
+    );
+    let public_key = private_key.public_key(&s).to_string();
+    let private_key = private_key.to_string();
+
+    Identity {
+        name,
+        company,
+        date_of_birth,
+        city_of_birth,
+        country_of_birth,
+        email,
+        postal_address,
+        public_key_pem,
+        private_key_pem,
+        bitcoin_public_key: public_key,
+        bitcoin_private_key: private_key.clone(),
+    }
+}
+
+pub fn write_identity_to_file(identity: &Identity) {
+    let data: Vec<u8> = identity_to_byte_array(identity);
+    fs::write(IDENTITY_FILE_PATH, data).expect("Unable to write file identity");
+}
+
+fn write_ed25519_keypair_to_file(ed25519_keys: &Keypair) {
+    let data: &[u8] = unsafe { util::structure_as_u8_slice(ed25519_keys) };
+    let data_sized = byte_array_to_size_array_keypair(data);
+    fs::write(IDENTITY_ED_25529_KEYS_FILE_PATH, *data_sized)
+        .expect("Unable to write keypair ed25519 in file");
+}
+
+fn write_peer_id_to_file(peer_id: &PeerId) {
+    let data: &[u8] = unsafe { util::structure_as_u8_slice(peer_id) };
+    let data_sized = byte_array_to_size_array_peer_id(data);
+    fs::write(IDENTITY_PEER_ID_FILE_PATH, *data_sized).expect("Unable to write peer id in file");
+}
+
+pub fn read_identity_from_file() -> Identity {
+    let data: Vec<u8> = fs::read(IDENTITY_FILE_PATH).expect("Unable to read file identity");
+    identity_from_byte_array(&data)
+}
+
+pub fn read_ed25519_keypair_from_file() -> Keypair {
+    let data: Vec<u8> =
+        fs::read(IDENTITY_ED_25529_KEYS_FILE_PATH).expect("Unable to read file keypair");
+    let key_pair_bytes_sized = byte_array_to_size_array_keypair(data.as_slice());
+    let key_pair: Keypair = unsafe { mem::transmute_copy(key_pair_bytes_sized) };
+    key_pair
+}
+
+pub fn read_peer_id_from_file() -> PeerId {
+    let data: Vec<u8> =
+        fs::read(IDENTITY_PEER_ID_FILE_PATH).expect("Unable to read file with peer id");
+    let peer_id_bytes_sized = byte_array_to_size_array_peer_id(data.as_slice());
+    let peer_id: PeerId = unsafe { mem::transmute_copy(peer_id_bytes_sized) };
+    peer_id
+}
+
+fn identity_to_byte_array(identity: &Identity) -> Vec<u8> {
+    to_vec(identity).unwrap()
+}
+
+fn identity_from_byte_array(identity: &[u8]) -> Identity {
+    Identity::try_from_slice(identity).unwrap()
+}
+
+pub fn byte_array_to_size_array_keypair(array: &[u8]) -> &[u8; size_of::<Keypair>()] {
+    array.try_into().expect("slice with incorrect length")
+}
+
+pub fn byte_array_to_size_array_peer_id(array: &[u8]) -> &[u8; size_of::<PeerId>()] {
+    array.try_into().expect("slice with incorrect length")
+}
