+++ src/external/mod.rs
@@ -1,2 +1,3 @@
 pub mod bitcoin;
+pub mod mint;
 pub mod time;
